ServerModel/requirements.txt
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
firebase-admin==6.5.0
gcloud==0.18.3
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
msgpack==1.0.8
oauth2client==4.1.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
tf-models-official==2.15.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
